# backend/src/main/java/com/backend/dto/CreateUserDto.java
package com.backend.dto;


import com.backend.model.domain.User;
import com.backend.model.enumaration.Role;

public record CreateUserDto(
        String username,
        String password,
        String repeatPassword,
        String name,
        String surname,
        Role role
) {

    /*
        todo: add repeat password logic
     */
    public User toUser() {
        return new User(username, password, name, surname, role);
    }
}